+++ CHANGELOG.md
@@ -8,14 +8,15 @@
 ### Removed
 
 ### Fixed
+
+- Actions error dialogs used message as title and title as message.
+- Don't fail "Manage Gradle Daemons" action if there is an I/O error during process collection.
+
 ## [1.2.3] - 2021-11-22
-### Added
 
 ### Changed
 
 - Fix incompatibilities with future IntelliJ releases
-
-### Removed
 
 ### Fixed
 ## [1.2.2] - 2021-08-29